sysinv/sysinv/sysinv/requirements.txt
pbr>=0.5
SQLAlchemy
amqplib>=0.6.1
anyjson>=0.3.3
argparse
eventlet==0.20.0
kombu>=2.4.8
lxml>=2.3
WebOb>=1.2.3,<1.3
greenlet>=0.3.2
sqlalchemy-migrate>=0.7
netaddr
paramiko>=1.8.0
passlib>=1.7.0
iso8601>=0.1.4
oslo.config>=3.7.0 # Apache-2.0
oslo.concurrency>=3.7.1 # Apache-2.0
oslo.db>=4.1.0 # Apache-2.0
oslo.service>=1.10.0 # Apache-2.0
oslo.utils>=3.5.0 # Apache-2.0
oslo.serialization<1.5.0
python-cinderclient==1.4.0
python-neutronclient>=2.2.3,<3
python-glanceclient>=0.9.0
python-keystoneclient>=0.3.0,<=2.0.0
keystonemiddleware==4.4.1
stevedore>=0.10
websockify>=0.5.1,<0.6
pecan>=0.2.0
six>=1.4.1
jsonpatch>=1.1
WSME>=0.5b2
Cheetah>=2.4.4
pyghmi
PyYAML>=3.10
python-magnumclient==2.3.1
psutil
